Player Health Regen
Player now regens health after a delay
Health regen stops upon taking damage

+++ FPS Raid Practive/Assets/ChargePanel.cs
@@ -6,9 +6,9 @@ using UnityEngine;
 public class ChargePanel : MonoBehaviour
 {
     public TextMeshPro ChargeLevelText;
-    public int StartingCharge = 0;
+    public float StartingCharge = 0;
 
-    private int _chargeLevel = 0;
+    private float _chargeLevel = 0;
 
     private void Start()
     {
@@ -19,7 +19,7 @@ public class ChargePanel : MonoBehaviour
         if (other.gameObject.CompareTag("Player")) 
         {
             PlayerController player = other.GetComponent<PlayerController>();
-            int chargeToAdd = player.PullHeldCharge();
+            float chargeToAdd = player.PullHeldCharge();
             _chargeLevel += chargeToAdd;
             ChargeLevelText.text = _chargeLevel.ToString();
         }
